💄 Add nameClearn util, fix pet name from overflow in the home cards
